fix: update metdata
